+++ run/kubejs/server_scripts/recipes/validation.js
@@ -38,4 +38,66 @@ ServerEvents.recipes(e => {
             }
         }
     });
+
+    var bool = true;
+    e.forEachRecipe({ tfc: {
+        type: 'is_tfc'
+    }}, r => bool = false);
+    if (bool) console.error('No TFC recipes?');
+    bool = true;
+    e.forEachRecipe({ tfc: {
+        type: 'has_isp'
+    }}, r => bool = false);
+    if (bool) console.error('No ISP recipes?');
+    bool = true;
+    e.forEachRecipe({ tfc: {
+        type: 'block',
+        block: 'tfc:rock/raw/dolomite'
+    }}, r => bool = false);
+    if (bool) console.error('No dolomite block ingredient');
+    bool = true;
+    /*
+    e.forEachRecipe({ tfc: {
+        type: 'fluid',
+        fluid: 'minecraft:water'
+    }}, r => false);
+    if (bool) console.error('No water ingredient');
+    bool = true;
+    */ // Nothing actually uses the basic fluid component
+    e.forEachRecipe({ tfc: {
+        type: 'fluid_stack',
+        fluid: Fluid.of('minecraft:milk', 20)
+    }}, r => bool = false);
+    if (bool) console.error('No milk ingredients');
+    bool = true;
+    e.forEachRecipe({ tfc: {
+        type: 'alloy_contents',
+        contents: 'tfc:copper'
+    }}, r => bool = false);
+    if (bool) console.error('No copper alloys');
+    bool = true;
+    e.forEachRecipe({ tfc: {
+        type: 'alloy_result',
+        result: 'tfc:rose_gold'
+    }}, r => bool = false);
+    if (bool) console.error('No rose gold alloying');
+    bool = true;
+    e.forEachRecipe({ tfc: {
+        type: 'isp',
+        match: 'tfc:powder/sulfur'
+    }}, r => bool = false);
+    if (bool) console.error('No sulfur ISPs');
+    bool = true;
+    e.forEachRecipe({ tfc: {
+        type: 'isp',
+        modifiers: 'tfc:copy_input'
+    }}, r => bool = false);
+    if (bool) console.error('No copy input ISPs');
+    bool = true;
+    e.forEachRecipe({ tfc: {
+        type: 'isp',
+        match: 'tfc:food/cooked_port',
+        modifiers: 'tfc:salted'
+    }}, r => bool = false);
+    if (bool) console.error('No salted pork ISPs');
 })