+++ changelog.md
@@ -5,6 +5,12 @@
 - Add noise inspection commands
 - Describing a knapping type via command will now include the texture(s) used by the knapping screen
 - Add support for Artisanal recipe types and ISMs
+- Fix `#notAxisAligned()` in ground cover builders creating invalid block state files
+- Custom fallen leaves are now included in the `tfc:fallen_leaves` block and item tags by default, making it able to snow pile and turn into hummus by default
+- Custom (fallen) leaves can now be removed from automatic tinting by calling `#noDynamicTinting()` on their builders
+- Tag ingredients print via the describe command can now be clicked to use the `/forge tags` command for that tag
+- Add TFC-specific recipe filters
+- Add metal map in misc sub-binding
 
 ### 1.3.2
 - Add `#hasPostProcess` and `#emissiveRendering` methods to extended properties